Adding Replication and persistence storage
Adding Replication and persistence storage
